% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agrega_mensal}
\alias{agrega_mensal}
\title{Agrega previsao mensal}
\usage{
agrega_mensal(simulacao, observacao)
}
\arguments{
\item{simulacao}{data table com a previsao contendo as seguintes colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}}

\item{observacao}{data table com o historico de vazao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da variavel}
}}
}
\value{
data.table com as colunas
\itemize{
\item{data_caso}{data da rodada}
\item{nome}{nome da sub-bacia}
\item{previsao}{ valor da previsao semanal}
\item{observacao}{valor observacao semanal}
\item{discretizacao}{discretizacao da previsao}
\item{horizonte}{horizonte da previsao}
}
}
\description{
Realiza agregacao mensal da previsao e observacao
}
