% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{analisa_previsoes}
\alias{analisa_previsoes}
\title{Analise das previsoes}
\usage{
analisa_previsoes(
  simulacao,
  observacao,
  semanal = TRUE,
  mensal = TRUE,
  anual = FALSE,
  inicio = NULL
)
}
\arguments{
\item{simulacao}{data table com a previsao contendo as seguintes colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}}

\item{observacao}{data table com o historico de vazao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da variavel}
}}

\item{semanal}{booleano indicando se a analise deve ser feita em acumulados semanais}

\item{mensal}{booleano indicando se a analise deve ser feita em acumulados semanais}

\item{anual}{booleano indicando se a analise deve ser feita em acumulados semanais}

\item{inicio}{}
}
\value{
saida lista contendo data table resultado com as colunas:
\itemize{
\item{nome}{nome da sub-bacia}
\item{metrica}{nome da metrica}
\item{valor}{valor da metrica}
\item{dia_previsao}{horizonte da previsao}
}
e data table resultado semanal com as colunas:
\itemize{
\item{nome}{nome da sub-bacia}
\item{metrica}{nome da metrica}
\item{valor}{valor da metrica}
\item{numero_semana}{horizonte semanal da previsao}
}
}
\description{
Calcula diversas metricas de avaliacao para series diaria, podendo fazer acumulados semanais, mensais,
sazonais, e anuais
}
