% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilacao.R
\name{assimilacao_oficial}
\alias{assimilacao_oficial}
\title{Assimilacao oficial}
\usage{
assimilacao_oficial(
  vetor_modelo,
  area,
  EbInic,
  TuInic,
  Supin,
  precipitacao,
  evapotranspiracao,
  evapotranspiracao_planicie,
  vazao,
  numero_dias,
  limite_prec = c(0.5, 2),
  limite_ebin = c(0.8, 1.2),
  limite_supin = c(0, 2),
  data_rodada
)
}
\arguments{
\item{vetor_modelo}{objeto de classe smap_ons}

\item{area}{area da sub-bacia}

\item{EbInic}{Escoamento Subterraneo inicial}

\item{TuInic}{Taxa de umidade inicial do solo}

\item{Supin}{Escoamento superficial inicial}

\item{precipitacao}{data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{evapotranspiracao}{vetor de evapotranspiracao}

\item{evapotranspiracao_planicie}{vetor de evapotranspiracao do reservatorio de planicie}

\item{vazao}{vetor de vazao observada}

\item{numero_dias}{numero de dias da assimilacao}

\item{limite_prec}{limites mínimo e máximo dos pesos utilizados para ponderar a precipitacao durante a etapa de assimilacao}

\item{limite_ebin}{limites mínimo e máximo da vazao de base inicial}

\item{limite_supin}{limites mínimo e máximo da vazao superficial inicial}

\item{data_rodada}{data da rodada}
}
\value{
ajuste lista contendo
\itemize{
\item{par}{parametros otimizados send}
\item{value}{valor da funcao objetivo utilizado}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}
}
\description{
Realiza a assimilacao de dados com a metodologia oficial
}
