% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcula_alfa}
\alias{calcula_alfa}
\title{Calcula alfa}
\usage{
calcula_alfa(
  simulacao,
  observacao,
  pesos = rep(1/length(observacao), length(observacao))
)
}
\arguments{
\item{simulacao}{vetor com os valores da serie simulada}

\item{observacao}{vetor com os valores da serie simulada}

\item{pesos}{pesos a serem utilizados para cada data}
}
\value{
correlacao ponderada entre as amostras
}
\description{
Realiza o calculo do termo alfa para o kge, representando a variabilidade dos erros
}
\examples{
observacao <- 1:30
simulacao <- observacao - 0.5
pbias <- calcula_alfa(simulacao, observacao)
}
