% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcula_dm}
\alias{calcula_dm}
\title{Realiza o calculo da DM}
\usage{
calcula_dm(
  simulacao,
  observacao,
  pesos = rep(1/length(observacao), length(observacao))
)
}
\arguments{
\item{simulacao}{vetor com os valores da serie simulada}

\item{observacao}{vetor com os valores da serie simulada}

\item{pesos}{pesos a serem utilizados para cada data}
}
\value{
dm distancia multicriterio
}
\description{
Realiza o calculo da DM
}
