% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcula_kge}
\alias{calcula_kge}
\title{Calcula kge}
\usage{
calcula_kge(
  simulacao,
  observacao,
  pesos = rep(1/length(observacao), length(observacao))
)
}
\arguments{
\item{simulacao}{vetor com os valores da serie simulada}

\item{observacao}{vetor com os valores da serie simulada}

\item{pesos}{pesos a serem utilizados para cada data}
}
\value{
Valor do KGE ponderado
}
\description{
Realiza o calculo do indice de eficiencia kge
}
\examples{
observacao <- 1:30
simulacao <- observacao - 0.5
kge <- calcula_kge(simulacao, observacao)
}
