% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcula_mape}
\alias{calcula_mape}
\title{Calcula MAPE}
\usage{
calcula_mape(
  simulacao,
  observacao,
  pesos = rep(1/length(observacao), length(observacao))
)
}
\arguments{
\item{simulacao}{vetor com os valores da serie simulada}

\item{observacao}{vetor com os valores da serie simulada}

\item{pesos}{pesos a serem utilizados para cada data}
}
\value{
mape
}
\description{
Realiza o calculo do MAPE
}
