% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcula_pbias}
\alias{calcula_pbias}
\title{Calcula PBIAS}
\usage{
calcula_pbias(
  simulacao,
  observacao,
  pesos = rep(1/length(observacao), length(observacao))
)
}
\arguments{
\item{simulacao}{vetor com os valores da serie simulada}

\item{observacao}{vetor com os valores da serie simulada}

\item{pesos}{pesos a serem utilizados para cada data}
}
\value{
vies percentual
}
\description{
Realiza o calculo do vies percentual de duas amostras
}
\examples{
observacao <- 1:30
simulacao <- observacao - 0.5
pbias <- calcula_pbias(simulacao, observacao)

}
