% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcula_rmse}
\alias{calcula_rmse}
\title{Calcula RMSE}
\usage{
calcula_rmse(
  simulacao,
  observacao,
  pesos = rep(1/length(observacao), length(observacao))
)
}
\arguments{
\item{simulacao}{vetor com os valores da serie simulada}

\item{observacao}{vetor com os valores da serie observada}

\item{pesos}{vetor de pesos a serem utilizados para cada data (default: pesos iguais)}
}
\value{
erro quadratico medio (RMSE)
}
\description{
Realiza o calculo da raiz do erro quadratico medio entre duas amostras
}
\examples{
observacao <- 1:30
simulacao <- observacao + rnorm(30, 0, 1)
rmse <- calcula_rmse(simulacao, observacao)

}
