% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibracao.R
\name{calibracao}
\alias{calibracao}
\title{funcao de calibracao do SMAP/ONS}
\usage{
calibracao(
  modelo,
  area,
  EbInic,
  TuInic,
  Supin,
  precipitacao,
  evapotranspiracao,
  vazao,
  data_inicio_objetivo,
  data_fim_objetivo,
  limite_inferior,
  limite_superior
)
}
\arguments{
\item{modelo}{vetor resultante de unlist do objeto de classe smap_ons}

\item{area}{area da sub-bacia}

\item{EbInic}{vazao de base inicial}

\item{TuInic}{umidade do solo inicial}

\item{Supin}{vazao superficial inicial}

\item{precipitacao}{data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{evapotranspiracao}{data table com a evapotranspiracao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{vazao}{data table com a vazao a avaliada
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{data_inicio_objetivo}{data inicial da avaliacao da funcao objetivo}

\item{data_fim_objetivo}{data final da avaliacao da funcao objetivo}

\item{limite_inferior}{vetor com o limite inferior dos parametros}

\item{limite_superior}{vetor com o limite superior dos parametros}
}
\value{
objetivo valor da funcao objetivo
}
\description{
funcao de calibracao do SMAP/ONS
}
