% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combina_observacao_previsao}
\alias{combina_observacao_previsao}
\title{Combina data table observado e previsto}
\usage{
combina_observacao_previsao(observado, previsto)
}
\arguments{
\item{observado}{data.table com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da serie_temporal observada}
}}

\item{previsto}{data.table com as colunas
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da previsao}
}}
}
\value{
serie_temporal data.table com as colunas
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da previsao}
}
}
\description{
Combina data table observado e previsto
}
