% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{completa_previsao}
\alias{completa_previsao}
\title{Completa as previsoes com mais n dias}
\usage{
completa_previsao(precipitacao_prevista, datas_rodadas, numero_dias = 2)
}
\arguments{
\item{precipitacao_prevista}{data.table com as colunas
\itemize{
\item{data_rodada - data da rodada}
\item{data_previsao - data da previsao}
\item{cenario - nome do cenario}
\item{nome - nome da sub-bacia}
\item{valor - valor da previsao}
}}

\item{datas_rodadas}{data table com as colunas
\itemize{
\item{data - data da rodada}
\item{numero_dias_previsao - numero de dias de previsão}
}}

\item{numero_dias}{numero de dias alem da data final da previsao a ser verificado}
}
\value{
precipitacao_prevista data.table com as colunas
\itemize{
\item{data_rodada - data da rodada}
\item{data_previsao - data da previsao}
\item{cenario - nome do cenario}
\item{nome - nome da sub-bacia}
\item{valor - valor da previsao}
}
}
\description{
Completa o data table de precisoes com mais N dias, para a ponderacao temporal ate t+2
}
