% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cria_datas}
\alias{cria_datas}
\title{Cria arquivo de datas a serem simuladas}
\usage{
cria_datas(data_inicio, data_fim)
}
\arguments{
\item{data_inicio}{data inicio da simulacao}

\item{data_fim}{data fim da simulacao}
}
\value{
data table contendo as datas dos casos a serem executados e seus respectivos horizontes:
\itemize{
\item{data - data do caso}
\item{numero_dias_previsao - horizonte do caso}
}
}
\description{
Cria arquivo de datas a serem simuladas
}
