% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cria_inicializacao}
\alias{cria_inicializacao}
\title{Cria arquivo de inicializacao}
\usage{
cria_inicializacao(parametros)
}
\arguments{
\item{parametros}{data.table com as colunas
\itemize{
\item{nome - nome da sub-bacia}
\item{parametro - nome do parametro}
\item{valor - valor do parametro}
}}
}
\value{
data.table com a inicializacao com as colunas
\itemize{
\item{nome - nome da sub_bacia}
\item{variavel - nome da variavel}
\item{valor - valor da variavel}
}
}
\description{
Cria arquivo de inicializacao
}
