% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibracao.R
\name{cria_kt}
\alias{cria_kt}
\title{Funcao de criacao de vetor de kts}
\usage{
cria_kt(kt_max, kt_min, alfa, beta)
}
\arguments{
\item{kt_max}{valor do maximo lag positivo}

\item{kt_min}{valor do maximo lag maximo negativo}

\item{alfa}{parametro alfa da distribuicao beta}

\item{beta}{parametro beta da distribuicao beta}
}
\value{
kt vetor de kts
}
\description{
funcao para a geracao do vetor de kt atraves de uma distribuicao beta
}
