% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cria_sub_bacias}
\alias{cria_sub_bacias}
\title{Cria arquivo de sub_bacias}
\usage{
cria_sub_bacias(parametros)
}
\arguments{
\item{parametros}{data.table com as colunas
\itemize{
\item{nome - nome da sub-bacia}
\item{parametro - nome do parametro}
\item{valor - valor do parametro}
}}
}
\value{
data.table sub_bacia com as colunas
\itemize{
\item{nome - nome da sub_bacia}
}
}
\description{
Cria arquivo de sub_bacias
}
