% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execucao.R
\name{executa_caso_oficial}
\alias{executa_caso_oficial}
\title{Executa caso oficial}
\usage{
executa_caso_oficial(pasta_caso)
}
\arguments{
\item{pasta_caso}{pasta do caso contendo as pastas "Arq_Entrada" e "Arq_Saida"}
}
\value{
Escreve os arquivos "sub-bacia_ajuste.txt" e "sub-bacia_cenario_previsao.txt" no formato oficial
}
\description{
executa rodada oficial do modleo SMAP/ONS
}
\examples{
\dontrun{
zip::unzip(system.file("extdata", "dados_entrada.zip", package = "smapOnsR"), exdir = system.file("extdata", package = "smapOnsR"))
pasta_entrada <- system.file("extdata", "Arq_Entrada1", package = "smapOnsR")
saida <- executa_caso_oficial(pasta_entrada)
}
}
