% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execucao.R
\name{executa_validacao_calibracao}
\alias{executa_validacao_calibracao}
\title{Executa caso oficial com entrada nova e aprimoramentos}
\usage{
executa_validacao_calibracao(pasta_caso)
}
\arguments{
\item{pasta_caso}{caminho da pasta contendo as pastas "Arq_Entrada" e
"Arq_Saida"}
}
\description{
executa rodada oficial do modelo SMAP/ONS com dados de entrada novos
e aprimoramentos
}
