% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizador.R
\name{executa_visualizador_calibracao}
\alias{executa_visualizador_calibracao}
\title{Visualizador de calibracao}
\usage{
executa_visualizador_calibracao()
}
\description{
Ferramenta de visualizacao de calibracao do modelo smap/ons
}
