% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizador_pmur.R
\name{executa_visualizador_calibracao_pmur}
\alias{executa_visualizador_calibracao_pmur}
\title{Visualizador de calibracao}
\usage{
executa_visualizador_calibracao_pmur()
}
\description{
Ferramenta de visualizacao de calibracao do modelo smap/ons
}
