% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizador.R
\name{executa_visualizador_previsao}
\alias{executa_visualizador_previsao}
\title{Visualizador SMAP/ONS}
\usage{
executa_visualizador_previsao()
}
\description{
Visualiza os cenarios gerados pelo modelo SMAP/ONS
}
