% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilacao.R
\name{funcao_objetivo_assimilacao}
\alias{funcao_objetivo_assimilacao}
\title{funcao objetivo de assimilacao do SMAP/ONS}
\usage{
funcao_objetivo_assimilacao(
  vetor_parametros,
  vetor_modelo,
  TuInic,
  precipitacao_ponderada,
  evapotranspiracao,
  evapotranspiracao_planicie,
  vazao,
  area,
  numero_dias,
  pesos_funcao_objetivo = rep(1/length(numero_dias), length(numero_dias))
)
}
\arguments{
\item{vetor_parametros}{vetor com os parametros:
\itemize{
\item{pesos}{pesos da ponderacao da precipitacao}
\item{Ebin}{vazao de base inicial}
\item{Supin}{vazao superficial inicial}
}}

\item{vetor_modelo}{vetor resultante de unlist do objeto de classe smap_ons}

\item{TuInic}{umidade do solo inicial}

\item{evapotranspiracao}{Vetor de ETo do final(ja corrigido e/ou ponderado)}

\item{vazao}{data table com a vazao a avaliada
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{area}{area da sub-bacia}

\item{numero_dias}{numero de dias da assimilacao}

\item{pesos_funcao_objetivo}{vetor de pesos da funcao objetivo}

\item{precipitacao}{Vetor de precipitacao final (ja corrigido e/ou ponderado)}
}
\value{
objetivo valor da funcao objetivo
}
\description{
funcao objetivo de assimilacao do SMAP/ONS
}
