% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibracao.R
\name{funcao_objetivo_calibracao}
\alias{funcao_objetivo_calibracao}
\title{funcao objetivo de calibracao do SMAP/ONS}
\usage{
funcao_objetivo_calibracao(
  vetor_modelo,
  kt_min,
  kt_max,
  area,
  EbInic,
  TuInic,
  Supin,
  precipitacao,
  evapotranspiracao,
  vazao,
  inicio_objetivo,
  fim_objetivo,
  postos_plu,
  pesos = rep(1/length(vazao[, valor]), length(vazao[, valor])),
  numero_dias,
  numero_postos_plu
)
}
\arguments{
\item{vetor_modelo}{vetor resultante de unlist do objeto de classe smap_ons}

\item{kt_min}{valor do maximo lag maximo negativo}

\item{kt_max}{valor do maximo lag positivo}

\item{area}{area da sub-bacia}

\item{EbInic}{vazao de base inicial}

\item{TuInic}{umidade do solo inicial}

\item{Supin}{vazao superficial inicial}

\item{precipitacao}{data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{evapotranspiracao}{data table com a evapotranspiracao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{vazao}{data table com a vazao a avaliada
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{inicio_objetivo}{inicio do calculo da funcao objetivo}

\item{fim_objetivo}{fim do calculo da funcao objetivo}

\item{postos_plu}{data.table com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{posto}{nome do posto plu}
\item{valor}{peso do posto plu}
}}

\item{pesos}{vetor de pesos a serem utilizados para cada data durante a calibracao}

\item{numero_dias}{numero de dias da simulacao para a calibracao}

\item{numero_postos_plu}{numero de posto_plu considerados}
}
\value{
objetivo valor da funcao objetivo
}
\description{
Realiza o calculo da funcao objetivo para a calibracao do SMAP/ONS
}
