% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historico_etp}
\alias{historico_etp}
\title{Historico de etp}
\format{
\subsection{\code{historico_etp}}{

A data table com 163,632 linhas and 4 colunas:
\describe{
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{prec}{valor de evapotranspiracao potencial observada}
}
}
}
}
\source{
ONS
}
\usage{
historico_etp
}
\description{
Historico de evapotranspiracao potencial proveniente do era 5 land
para o periodo de 1991 a 2022
}
\keyword{datasets}
