% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historico_etp_NC}
\alias{historico_etp_NC}
\title{Historico de Normais Climatologicas de etp}
\format{
\subsection{\code{historico_etp}}{

Data table com 1,320 linhas and 4 colunas:
\describe{
\itemize{
\item{mes}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor de evapotranspiracao potencial observada}
}
}
}
}
\source{
ONS
}
\usage{
historico_etp_NC
}
\description{
Historico de Normais Climatologicas de de evapotranspiracao potencial
utilizados oficialmente no SMAP/ONS
}
\keyword{datasets}
