% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historico_precipitacao}
\alias{historico_precipitacao}
\title{Historico de PSAT}
\format{
\subsection{\code{historico_precipitacao}}{

A data table com 552,258 linhas and 4 colunas:
\describe{
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor de precipitacao observada}
}
}
}
}
\source{
ONS
}
\usage{
historico_precipitacao
}
\description{
Historico de precipitação das sub-bacias modeladas pelo SMAP/ONS
para o periodo de 2010 a 2022
}
\keyword{datasets}
