% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historico_vazao}
\alias{historico_vazao}
\title{Historico de vazao}
\format{
\subsection{\code{historico_vazao}}{

A data table com 486,513 linhas and 4 colunas:
\describe{
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{valor - valor de vazao observada}
}
}
}
}
\source{
ONS
}
\usage{
historico_vazao
}
\description{
Historico de vazao das sub-bacias modeladas pelo SMAP/ONS
para o periodo de 2010 a 2021
}
\keyword{datasets}
