% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicializacao.R
\name{inicializacao_smap}
\alias{inicializacao_smap}
\title{Funcao de calculo dos valores inicias}
\usage{
inicializacao_smap(
  vetor_modelo,
  area,
  EbInic = 0,
  TuInic = 0.3,
  Supin = 100,
  Rsup2Inic = 0
)
}
\arguments{
\item{vetor_modelo}{vetor resultante unlist do objeto de classe smap_ons}

\item{area}{area da sub-bacia}

\item{EbInic}{Escoamento Subterraneo inicial}

\item{TuInic}{Taxa de umidade inicial do solo}

\item{Supin}{Escoamento superficial inicial}

\item{Rsup2Inic}{Escoamento superficial inicial referente ao reservatorio de planicie}
}
\value{
List com os parametros:
\itemize{
\item{EbInic}{vazao de base inicial}
\item{TuInic}{percentual de umidade do solo inicial}
\item{Supin}{vazao superficial inicial}
\item{Rsup2Inic}{valor do reservatorio de planicie inicial}
\item{RsoloInic}{valor do reservatorio de solo inicial}
\item{RsupInic}{valor do reservatorio superficial inicial}
\item{RsubInic}{valor do reservatorio subterraneo inicial}
}
}
\description{
Realiza o calculo dos Valores iniciais para rodada do SMAP/ONS
}
\details{
Funcao para criar objeto com os valores iniciais de variaveis de estado para rodar o SMAP
}
