% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_saida.R
\name{le_ajuste}
\alias{le_ajuste}
\title{Le Arquivos _AJUSTE}
\usage{
le_ajuste(arq, data, subbacia)
}
\arguments{
\item{arq}{arquivo AJUSTE a ser lido}

\item{data}{data de referencia da rodada. Se omitido, sera usada a maior data no arquivo + 1}

\item{subbacia}{nome da subbacia sendo lida. Se omitido sera inferido do nome do arquivo}
}
\value{
data.table do arquivo arrumado
\itemize{
\item{dia_assimilacao} numero de dias atras da variavel assimilada com respeito a rodada
\item{tu} variavel TU da assimilacao
\item{escoamento_base} escoamento de base da assimilacao
\item{escoamento_superficial} escoamento superficial da assimilacao
\item{vazao_calculada} vazao calculada da assimilacao
\item{data_previsao} dia referencia de execucao da previsao
\item{subbacia} nome da subbacia da qual estes dados dizem respeito
}
}
\description{
Leitor de arquivos _AJUSTE da assimilacao de dados
}
