% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_arq_entrada}
\alias{le_arq_entrada}
\title{Leitor de arquivo de previsao de precipitacao do smap/ons}
\usage{
le_arq_entrada(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}
}
\value{
saida lista com os data.tables: vazao
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{valor}{valor da previsao}
}
}
\description{
Le arquivo "'nome_cenario'\emph{p}'data_caso'_'data_previsao'.txt" utilizado no aplicativo SMAP/ONS
}
