% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_arq_entrada_novo}
\alias{le_arq_entrada_novo}
\title{Le arquivos de entrada em novo formato}
\usage{
le_arq_entrada_novo(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{nome da pasta com os arquivos de entrada}
}
\value{
saida lista contendo os seguintes data tables:
}
\description{
Realiza a leitura dos arquivos de entrada em novo formato
}
