% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_arquivos}
\alias{le_arquivos}
\title{Le csv com o nome dos arquivos do caso}
\usage{
le_arquivos(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{caminho da pasta com os arquivos de entrada}
}
\value{
data table contendo as datas dos casos a serem executados e seus respectivos horizontes:
\itemize{
\item{data - data do caso}
\item{numero_dias_previsao - horizonte do caso}
}
}
\description{
Realiza a leitura do csv 'arquivos.csv' com os arquivos a serem lidos
}
