% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_nova.R
\name{le_datas_calibracao}
\alias{le_datas_calibracao}
\title{le_datas_calibracao}
\usage{
le_datas_calibracao(arq)
}
\arguments{
\item{arq}{o arquivo do tipo "sub-bacia_PARAMETROS.txt"}
}
\value{
data.table com as colunas
\itemize{
\item{nome - nome da sub-bacia}
\item{parametro - nome do parametro}
\item{valor - valor do parametro}
}
}
\description{
Leitor de arquivo de parametros para obtencao das datas utilizadas
na calibracao
}
\details{
Le arquivo de parametros "sub-bacia_PARAMETROS.txt".
}
