% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_datas_rodada}
\alias{le_datas_rodada}
\title{Le csv com dados de datas de execucao do caso}
\usage{
le_datas_rodada(arq)
}
\arguments{
\item{arq}{nome do arquivo "inicializacao.csv"}
}
\value{
data.table evapotranspiracao com as colunas
\itemize{
\item{mes}{mes da NC}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da NC de evapotranspiracao observada}
}
}
\description{
Realiza a leitura do csv 'datasRodadas.csv' com as datas e o horizonte de previsao
}
