% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_antiga.R
\name{le_entrada_bat_conf}
\alias{le_entrada_bat_conf}
\title{Le arquivo bat.conf}
\usage{
le_entrada_bat_conf(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{o arquivo do tipo "caso.txt"}
}
\value{
formato_previsao numero do tipo de arquivo de previsao
}
\description{
Leitor do arquivo de entrada bat.txt contendo para receber qual padrao de arquivo de previsao esta
}
