% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_antiga.R
\name{le_entrada_caso}
\alias{le_entrada_caso}
\title{Le arquivo caso.txt}
\usage{
le_entrada_caso(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{o arquivo do tipo "caso.txt"}
}
\value{
caso lista contendo os seguintes parametros
\itemize{
\item{numero_subbacias- numero de sub-bacias do caso}
\item{nome_subbacia - vetor com o nome das sub-bacias}
}
}
\description{
Leitor do arquivo de entrada caso.txt contendo quais sub-bacias constam no caso
}
