% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_evapotranspiracao}
\alias{le_entrada_evapotranspiracao}
\title{Leitor de arquivo de normal climatologica de evapotranspiracao}
\usage{
le_entrada_evapotranspiracao(pasta_entrada, nome_subbacia)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}
}
\value{
data.table evapotranspiracao com as colunas
\itemize{
\item{mes}{mes da NC}
\item{posto}{nome do posto}
\item{valor}{valor da NC de evapotranspiracao observada}
}
}
\description{
Le arquivo evapotranspiracao utilizado no aplicativo SMAP/ONS
}
