% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_inicializacao}
\alias{le_entrada_inicializacao}
\title{Leitor de arquivo de inicializacao do smap}
\usage{
le_entrada_inicializacao(pasta_entrada, nome_subbacia)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}
}
\value{
data.table entrada_inicializacao com as colunas
\itemize{
\item{valor}{valor do parametro}
\item{parametro}{nome do parametros}
}
}
\description{
Le arquivo "sub-bacia_inicializacao.txt" utilizado no aplicativo SMAP/ONS
}
