% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_antiga.R
\name{le_entrada_modelos_precipitacao}
\alias{le_entrada_modelos_precipitacao}
\title{Leitor de arquivo de modelos de previsao de precipitacao do smap/ons}
\usage{
le_entrada_modelos_precipitacao(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}
}
\value{
modelos_precipitacao lista contendo os seguintes parametros
\itemize{
\item{numero_cenario - numero de cenarios do caso}
\item{nome_cenario - data table contendo os seguintes parametros}
\itemize{
\item{nome_cenario_completo - nome completo dos cenarios de precipitacao considerados o caso}
\item{nome_cenario_1 - primeira parte do nome dos cenarios de precipitacao considerados o caso}
\item{nome_cenario_2 - segunda parte do nome dos cenarios de precipitacao considerados o caso}
}
}
}
\description{
Le arquivo "modelos_precipitacao.txt" utilizado no aplicativo SMAP/ONS
}
