% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_pontos_grade}
\alias{le_entrada_pontos_grade}
\title{Leitor de arquivo de com os pontos de grade de previsao da sub_bacia}
\usage{
le_entrada_pontos_grade(pasta_entrada, nome_subbacia, modelos_precipitacao)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}

\item{modelos_precipitacao}{lista contendo os seguintes parametros
\itemize{
\item{nome_cenario}{nome dos cenarios de precipitacao considerados o caso}
\item{numero_cenario}{numero de cenarios do caso}
}}
}
\value{
data.table pontos_grade com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{nome_cenario}{nome nome dos cenarios de precipitacao considerados o caso}
\item{latitude}{latitude do ponto de grade do cenário}
\item{longitude}{longitude do ponto de grade do cenário}
}
}
\description{
Le arquivo "sub-bacia_nome_cenario.txt" utilizado no aplicativo SMAP/ONS
}
