% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_antiga.R
\name{le_entrada_posto_plu}
\alias{le_entrada_posto_plu}
\title{Leitor de arquivo com os postos plu da sub_bacia}
\usage{
le_entrada_posto_plu(pasta_entrada, nome_subbacia)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}
}
\value{
data.table postos_plu com as colunas
\itemize{
\item{nome - nome da sub_bacia}
\item{posto - nome do posto plu}
\item{valor - peso do posto plu}
}
}
\description{
Le arquivo "sub-bacia_postos_plu.txt" utilizado no aplicativo SMAP/ONS
}
