% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_precipitadao}
\alias{le_entrada_precipitadao}
\title{Leitor de arquivo de precipitacao observada do smap/ons}
\usage{
le_entrada_precipitadao(pasta_entrada, postos_plu)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{posto_plus}{data.table postos_plu com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{posto}{nome do posto plu}
\item{valor}{peso do posto plu}
}}
}
\value{
data.table vazao com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da precipitacao observada}
}
}
\description{
Le arquivo "nome_posto_plu_c.txt" utilizado no aplicativo SMAP/ONS
}
