% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_previsao_precipitacao_0}
\alias{le_entrada_previsao_precipitacao_0}
\title{Leitor de arquivo de previsao de precipitacao do smap/ons}
\usage{
le_entrada_previsao_precipitacao_0(
  pasta_entrada,
  data_caso,
  data_previsao,
  pontos_grade
)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{posto_plus}{data.table postos_plu com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{posto}{nome do posto plu}
\item{valor}{peso do posto plu}
}}
}
\value{
data.table vazao com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da precipitacao observada}
}
}
\description{
Le arquivo "'nome_cenario'\emph{p}'data_caso'_'data_previsao'.txt" utilizado no aplicativo SMAP/ONS
}
