% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_vazao}
\alias{le_entrada_vazao}
\title{Leitor de arquivo de vazao observada do smap/ons}
\usage{
le_entrada_vazao(pasta_entrada, nome_subbacia)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}
}
\value{
data.table vazao com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da precipitacao observada}
}
}
\description{
Le arquivo "sub-bacia_inicializacao.txt" utilizado no aplicativo SMAP/ONS
}
