% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_evapotranspiracao}
\alias{le_evapotranspiracao}
\title{Leitor de arquivo de normal climatologica de evapotranspiracao}
\usage{
le_evapotranspiracao(arq)
}
\arguments{
\item{arq}{o arquivo do tipo "subbacia_EVAPOTRANSPIRACAO.txt"}
}
\value{
data.table evapotranspiracao com as colunas
\itemize{
\item{mes}{mes da NC}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da NC de evapotranspiracao observada}
}
}
\description{
Le arquivo evapotranspiracao utilizado no aplicativo SMAP/ONS
}
