% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_evapotranspiracao_nc}
\alias{le_evapotranspiracao_nc}
\title{Leitura de csv com NC de evapotranspiracao}
\usage{
le_evapotranspiracao_nc(arq)
}
\arguments{
\item{arq}{nome do arquivo contendo a relacao de postos plu x sub-bacia}
}
\value{
data.table evapotranspiracao com as colunas
\itemize{
\item{mes}{mes da NC}
\item{posto}{nome do posto}
\item{valor}{valor da NC de evapotranspiracao observada}
}
}
\description{
Le arquivo csv contendo as normais climatologias evapotranspiracao utilizado no aplicativo SMAP/ONS
}
