% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_saida.R
\name{le_execucao}
\alias{le_execucao}
\title{Le Arquivos execucao}
\usage{
le_execucao(pasta_saida, data)
}
\arguments{
\item{pasta_saida}{pasta com os arquivos de saida}

\item{data}{data de referencia da rodada. Se omitido, sera inferido pelo nome do arquivo}
}
\value{
Lista de dois data.tables: um de assimilacao e outro de previsao. O primeiro de
assimilacao:
\itemize{
\item{subbacia} nome da subbacia da qual estes dados dizem respeito
\item{data} dia referencia de execucao da assimilacao
\item{dia_assimilacao} numero de dias atras da variavel assimilada com respeito a rodada
\item{qcal} vazao calculada da assimilacao
\item{rsolo} variavel Rsolo da assimilacao
\item{rsub} variavel Rsub da assimilacao
\item{rsup} variavel Rsup da assimilacao
\item{peso_chuva} pesos das precs passadas na assimilacao
\item{ebin} valor de ebin obtido na assimilacao
\item{supin} valor de supin obtido na assimilacao
\item{funcao_objetivo} valor da funcao objetivo obtida (valores repetidos por subbacia)
\item{data_caso} data do caso
}
}
\description{
Leitor dos arquivos execucao da assimilacao de dados
}
