% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_inicializacao}
\alias{le_inicializacao}
\title{Le csv com dados de inicializacao do caso}
\usage{
le_inicializacao(arq)
}
\arguments{
\item{arq}{nome do arquivo "inicializacao.csv"}
}
\value{
data.table com a inicializacao com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{Ebin}{vazao de base inicial}
\item{Supin}{vazao superficial inicial}
\item{Yuin}{umidade do solo inicial}
}
}
\description{
Realiza a leitura do csv 'inicializacao.csv' com os dados iniciais
}
