% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_parametros}
\alias{le_parametros}
\title{le_parametros}
\usage{
le_parametros(arq)
}
\arguments{
\item{arq}{o arquivo do tipo "sub-bacia_PARAMETROS.txt"}
}
\value{
data.table com as colunas
\itemize{
\item{Nome}{Nome da sub-bacia}
\item{parametro}{nome do parametro}
\item{valor}{valor do parametro}
}
}
\description{
Leitor de arquivo de parametros
}
\details{
Le arquivo de parametros "sub-bacia_PARAMAETROS.txt".
}
