% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_postos_plu}
\alias{le_postos_plu}
\title{Le csv com a relaca de postos plu x sub_bacia}
\usage{
le_postos_plu(arq)
}
\arguments{
\item{arq}{nome do arquivo contendo a relacao de postos plu x sub-bacia}
}
\value{
data table contendo as datas dos casos a serem executados e seus respectivos horizontes:
\itemize{
\item{data}{data do caso}
\item{numero_dias_previsao}{horizonte do caso}
}
}
\description{
Realiza a leitura do csv 'postos_plu.csv' csv com a relacao de postos plu x sub_bacia
}
