% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_saida.R
\name{le_previsao}
\alias{le_previsao}
\title{Le Arquivos PREVISAO}
\usage{
le_previsao(pasta_saida, data)
}
\arguments{
\item{pasta_saida}{pasta com os arquivos de saida}

\item{data}{data de referencia da rodada. Se omitido, sera inferido pelo nome do arquivo}
}
\value{
data.table contendo a dado de previsao arrumado
\itemize{
\item{data_caso} data do caso executado
\item{data_previsao} dia alvo da previsao
\item{cenario} cenario de precipitacao utilizado
\item{nome} nome da subbacia da qual estes dados dizem respeito
\item{variavel} nome da variavel prevista
\item{valor} valor da variavel prevista
}
}
\description{
Leitor dos arquivos de previsao de cada rodada diaria
}
