% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_ons.R
\name{new_modelo_smap_ons_pmur}
\alias{new_modelo_smap_ons_pmur}
\title{Construtor do modelo classe \code{smap_ons}}
\usage{
new_modelo_smap_ons_pmur(parametros, postos_plu)
}
\arguments{
\item{parametros}{data table com 8910 linhas and 3 colunas:
\describe{
\itemize{
\item{nome - nome da sub-bacia}
\item{parametros - nome do parametros}
\item{valor - valor do parametro}
}
}}

\item{postos_plu}{data.table  com as colunas
\itemize{
\item{nome - nome da sub_bacia}
\item{posto - nome do posto plu}
\item{valor - peso do posto plu}
}}
}
\value{
objeto de classe \code{smap_ons}
}
\description{
Construtor do modelo classe \code{smap_ons}
}
