% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordem_afluencia}
\alias{ordem_afluencia}
\title{Ordem de afluencia}
\usage{
ordem_afluencia(montante, jusante)
}
\arguments{
\item{montante}{Nomes ou IDs das UHEs de montante da cascata}

\item{jusante}{Nomes ou IDs das UHEs imediatamente a jusante de cada UHE}
}
\value{
Retorna a ordem de afluencia para cada UHE
}
\description{
Ordem de afluencia
}
