% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parametros}
\alias{parametros}
\title{Parametros}
\format{
\subsection{\code{parametros}}{

A data table com 110 linha and 83 colunas:
\describe{
\itemize{
\item{nome}{nome da sub-bacia}
\item{area}{area da sub-bacia}
\item{nKt}{número de kts}
\item{kt2}{valor do parametro kt+2}
\item{kt-60}{valor do parametro kt-60}
}
}
}
}
\source{
ONS
}
\usage{
parametros
}
\description{
Parametros das sub-bacias modeladas pelo SMAP/ONS
}
\keyword{datasets}
