% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{poderacao_temporal}
\alias{poderacao_temporal}
\title{Realiza a ponderacao temporal de variaveis}
\usage{
poderacao_temporal(
  serie_temporal,
  pesos,
  kt_min,
  kt_max,
  data_inicio,
  data_fim
)
}
\arguments{
\item{serie_temporal}{data table com a variavel a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{pesos}{Vetor de pesos iniciando em kt+2 ate kt-60}

\item{kt_min}{numero de dias para trás a serem considerados na ponderacao}

\item{kt_max}{numero de dias a frente a serem considerados na ponderacao}

\item{data_inicio}{data de inicio da construcao da serie temporal ponderada}

\item{data_fim}{data de inicio da construcao da serie temporal ponderada}
}
\value{
serie_temporal_ponderada: data table com a variavel ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel ponderada}
}
}
\description{
Realiza a ponderacao temporal de variaveis
}
