% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{ponderacao_espacial}
\alias{ponderacao_espacial}
\title{Funcao de ponderacao espacial}
\usage{
ponderacao_espacial(historico_precipitacao, postos_plu)
}
\arguments{
\item{historico_precipitacao}{data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto plu}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{postos_plu}{data.table postos_plu com as colunas
\itemize{
\item{posto}{nome da sub_bacia}
\item{psat}{nome do posto plu}
\item{valor}{peso do posto plu}
}}
}
\value{
precipitacao data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{sub_bacia}{nome da sub_bacia}
\item{id}{id da sub_bacia}
\item{valor}{valor da precipitacao ponderada}
}
}
\description{
Realiza a ponderacao espacial de precipitacao atraves dos pesos de cada sub_bacia
}
