% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{ponderacao_espacial_previsao}
\alias{ponderacao_espacial_previsao}
\title{Funcao de ponderacao espacial}
\usage{
ponderacao_espacial_previsao(precipitacao_prevista, postos_plu)
}
\arguments{
\item{precipitacao_prevista}{data.table com a precipitacao prevista por posto pluviometrico com as colunas
\itemize{
\item{data_rodada - data da rodada do modelo que gerou a previsao}
\item{data_previsao - data da previsao}
\item{cenario - codigo do cenario}
\item{posto - nome do posto pluviometrico}
\item{valor - valor da previsao de precipitacao}
}}

\item{postos_plu}{data.table postos_plu com as colunas
\itemize{
\item{posto}{nome da sub_bacia}
\item{psat}{nome do posto plu}
\item{valor}{peso do posto plu}
}}
}
\value{
data.table com a precipitacao prevista por sub-bacia com as colunas
\itemize{
\item{data_rodada - data da rodada do modelo que gerou a previsao}
\item{data_previsao - data da previsao}
\item{cenario - codigo do cenario}
\item{nome - nome da sub-bacia}
\item{valor - valor da previsao de precipitacao}
}
}
\description{
Realiza a ponderacao espacial de previsao de precipitacao atraves dos pesos de cada sub_bacia
}
