% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{ponderacao_temporal}
\alias{ponderacao_temporal}
\title{Realiza a ponderacao temporal de variaveis}
\usage{
ponderacao_temporal(serie_temporal, kt, kt_max, kt_min)
}
\arguments{
\item{serie_temporal}{vetor com a variavel a ser ponderada}

\item{kt}{vetor de kts}

\item{kt_max}{valor do maximo lag positivo}

\item{kt_min}{valor do maximo lag maximo negativo}
}
\value{
serie_temporal_ponderada: data table com a variavel ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel ponderada}
}
}
\description{
Realiza a ponderacao temporal de variaveis
}
