% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{ponderacao_temporal2}
\alias{ponderacao_temporal2}
\title{Realiza a ponderacao temporal de variaveis}
\usage{
ponderacao_temporal2(serie_temporal, kt)
}
\arguments{
\item{serie_temporal}{vetor com a variavel a ser ponderada}

\item{kt}{vetor de kts}
}
\value{
serie_temporal_ponderada: data table com a variavel ponderada com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel ponderada}
}
}
\description{
Realiza a ponderacao temporal de variaveis
}
