% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postos_plu}
\alias{postos_plu}
\title{Postos plu}
\format{
\subsection{'postos_plu'}{

data table com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{posto}{nome do posto plu}
\item{valor}{peso do posto plu}
}
}
}
\source{
ONS
}
\usage{
postos_plu
}
\description{
Relacao de posto plu por sub-bacia
}
\keyword{datasets}
