% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodadas_encadeadas.R
\name{rodada_encadeada_etp}
\alias{rodada_encadeada_etp}
\title{Funcao para realizar rodadas encadeadas considerando etp variavel do SMAP/ONS}
\usage{
rodada_encadeada_etp(
  parametros,
  inicializacao,
  historico_precipitacao,
  previsao_precipitacao,
  historico_etp,
  previsao_etp,
  historico_vazao,
  postos_plu,
  datas_rodadas,
  numero_cenarios,
  sub_bacias
)
}
\arguments{
\item{parametros}{data table com os parametros dos modelos
\itemize{
\item{Nome}{Nome da sub-bacia}
\item{parametro}{nome do parametro}
\item{valor}{valor do parametro}
}}

\item{inicializacao}{data.table com a inicializacao com as colunas:
\itemize{
\item{nome}{nome da sub_bacia}
\item{variavel}{vazao de base inicial}
\item{valor}{valor da variavel}
}}

\item{historico_precipitacao}{data table com o historico de precipitacao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da variavel}
}}

\item{previsao_precipitacao}{data.table com a previsao de precipitacao com as colunas:
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{posto}{nome do posto}
\item{valor}{valor da previsao}
}}

\item{historico_etp}{data.table com o historico de evapotranspiracao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da variavel}
}}

\item{previsao_etp}{data.table com a previsao de evapotranspiracao com as colunas:
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{posto}{nome do posto}
\item{valor}{valor da previsao}
}}

\item{historico_vazao}{data table com o historico de vazao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da variavel}
}}

\item{postos_plu}{data table contendo a relacao sub-bacia x postos_plu com as colunas:
\itemize{
\item{nome}{nome da sub_bacia}
\item{posto}{nome do posto plu}
\item{valor}{peso do posto plu}
}}

\item{datas_rodadas}{data table contendo as datas dos casos a serem executados e seus respectivos horizontes:
\itemize{
\item{data}{data do caso}
\item{numero_dias_previsao}{horizonte do caso}
}}

\item{numero_cenarios}{numero de cenarios a serem gerados}

\item{sub_bacias}{vetor com o nome das sub-bacias a serem consideradas}
}
\value{
saida lista com o data.table previsao contendo as seguintes colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
otimizacao data table com as colunas:
\itemize{
\item{otimizacao}{valor das variaveis da otimizacao}
\item{nome}{nome da sub-bacia}
\item{data_caso}{data do caso}
}
assimilacao data table com as colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_assimilacao}{data da assimilacao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
precipitacao data table contendo precipitacao observada e previsata com as colunas:
\itemize{
\item{data_previsao}{data da precipitacao}
\item{data_rodada}{data da rodada}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
}
\description{
Realiza execucoes encadeadas do modelo SMAP/ONS considerando etp variavel
}
