% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_ons.R
\name{rodada_varios_dias}
\alias{rodada_varios_dias}
\title{Rodada de varios dias do SMAP}
\usage{
rodada_varios_dias(
  modelo,
  inicializacao,
  precipitacao,
  evapotranspiracao,
  Emarg,
  numero_dias
)
}
\arguments{
\item{modelo}{objeto de classe smap_ons}

\item{inicializacao}{Objeto resultante da funcao smap.inicializacao}

\item{precipitacao}{Vetor de precipitacao final (ja corrigido e/ou ponderado)}

\item{evapotranspiracao}{Vetor de ETo do final(ja corrigido e/ou ponderado)}

\item{Emarg}{Vetor da evaporacao final do reservatorio de planicie (ja corrigido e/ou ponderado)}

\item{numero_dias}{numero de dias da rodada}
}
\value{
Matriz com a vazao calculada e os valores de estado e funcoes de transferencia
}
\description{
Rodada de varios dias do SMAP
}
