% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicializacao.R
\name{smap_ons.inic}
\alias{smap_ons.inic}
\title{Valores iniciais SMAP}
\usage{
smap_ons.inic(parametros, EbInic = 0, TuInic = 0.3, Supin = 100, Rsup2Inic = 0)
}
\arguments{
\item{EbInic}{Escoamento Subterraneo inicial}

\item{TuInic}{Taxa de umidade inicial do solo}

\item{Supin}{Escoamento superficial inicial}

\item{Rsup2Inic}{Escoamento superficial inicial referente ao reservatorio de planicie}
}
\value{
List com os parametros
}
\description{
Funcao para criar objeto com os valores iniciais de variaveis de estado para rodar o SMAP
}
