% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_ons.R
\name{smap_ons.previsao}
\alias{smap_ons.previsao}
\title{Rodada de um dia do SMAP}
\usage{
smap_ons.previsao(
  parametrosSMAP,
  inicializacao,
  precipitacao,
  evapotranspiracao,
  Emarg,
  mRbind,
  dataD = NULL
)
}
\arguments{
\item{parametrosSMAP}{data.table contendo n linhas and 83 colunas:
\describe{
\itemize{
\item{nome}{nome da sub-bacia}
\item{area}{area da sub-bacia}
\item{nKt}{número de kts}
\item{kt2}{valor do parametro kt+2}
\item{kt-60}{valor do parametro kt-60}
}
}}

\item{inicializacao}{Objeto resultante da funcao smap.inicializacao}

\item{precipitacao}{Valor de precipitacao final do dia (ja corrigido e/ou ponderado)}

\item{evapotranspiracao}{Valor de ETo do final dia (ja corrigido e/ou ponderado)}

\item{Emarg}{Valor da evaporacao final do dia do reservatorio de planicie (ja corrigido e/ou ponderado)}

\item{mRbind}{Matriz com resultado da iteracao passada (primeira iteracao nao atribuir nenhum valor)}

\item{dataD}{Nomea a linha da matriz de saida com a data fornecida}
}
\value{
Matriz com a vazao calculada e os valores de estado e funcoes de transferencia
}
\description{
Rodada de um dia do SMAP
}
