% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaliza.R
\name{totaliza_previsao}
\alias{totaliza_previsao}
\title{Totaliza previsoes}
\usage{
totaliza_previsao(previsao, vazao_observada, configuracao)
}
\arguments{
\item{previsao}{data.table previsao contendo as seguintes colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}}

\item{vazao_observada}{data.table com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da precipitacao observada}
}}

\item{configuracao}{data.table com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{valor}{valor da precipitacao observada}
}}
}
\value{
previsao_totalizada - data.table contendo as seguintes colunas
\itemize{
\item{data_rodada - data em que o caso foi executado}
\item{data_previsao - data da previsao realizada}
\item{nome - nome da UHE}
\item{cenario - nome do cenario}
\item{previsao_distribuida}{valor da vazao apos o processo de distribuicao de incrementais agrupadas previstas}
\item{previsao_incremental}{valor da vazao incremental prevista}
\item{previsao_total}{valor da vazao total prevista, considerando o tempo de viagem entre as UHES}
\item{previsao_total_sem_tv}{valor da vazao total prevista, sem copnsiderar o tempo de viagem entre as UHES}
}
}
\description{
Funcao para totalizar as previsoes do SMAP/ONS, atraves de metodos de propagacao
}
