% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transforma_historico_previsao}
\alias{transforma_historico_previsao}
\title{Transforma historico de serie temporal em data table de previsao}
\usage{
transforma_historico_previsao(serie_temporal, datas_rodadas)
}
\arguments{
\item{serie_temporal}{data.table com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da serie_temporal observada}
}}

\item{datas_rodadas}{data.table com as colunas
\itemize{
\item{data}{data da rodada}
\item{numero_dias_previsao}{numero de dias de previsao}
}}
}
\value{
data.table com as colunas
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{valor}{valor da previsao}
}
}
\description{
Transforma historico de serie temporal em data table de previsao
}
