% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao_entrada.R
\name{valida_cenarios}
\alias{valida_cenarios}
\title{valida_cenarios}
\usage{
valida_cenarios(evapotranspiracao_prevista, precipitacao_prevista)
}
\arguments{
\item{evapotranspiracao_prevista}{data.table com a evapotranspiracao prevista com as colunas
\itemize{
\item{data_rodada - data da rodada do modelo que gerou a previsao}
\item{data_previsao - data da previsao}
\item{cenario - codigo do cenario}
\item{nome - nome da sub bacia}
\item{valor - valor da previsao de precipitacao}
}}

\item{precipitacao_prevista}{data.table com a precipitacao prevista com as colunas
\itemize{
\item{data_rodada - data da rodada do modelo que gerou a previsao}
\item{data_previsao - data da previsao}
\item{cenario - codigo do cenario}
\item{nome - nome da sub bacia}
\item{valor - valor da previsao de precipitacao}
}}
}
\description{
Valida se para cada valor unico de nome, data_previsao e data_rodada existem os mesmos
cenarios
}
