% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao_entrada.R
\name{valida_duplicadas}
\alias{valida_duplicadas}
\title{valida_duplicadas}
\usage{
valida_duplicadas(dt, colunas_chave, nome_dt)
}
\arguments{
\item{dt}{data.table com a precipitacao prevista com as colunas
\itemize{
\item{data_rodada - data da rodada do modelo que gerou a previsao}
\item{data_previsao - data da previsao}
\item{cenario - codigo do cenario}
\item{nome - nome da sub bacia}
\item{valor - valor da previsao de precipitacao}
}}

\item{colunas_chave}{vetor com os nomes das colunas que compoem a chave unica}

\item{nome_dt}{nome do data.table (string) para usar na mensagem de erro}
}
\description{
Valida se existe linhas duplicadas no data.table
}
