% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao_entrada.R
\name{valida_previsao_prec}
\alias{valida_previsao_prec}
\title{valida_previsao_prec}
\usage{
valida_previsao_prec(precipitacao_prevista)
}
\arguments{
\item{precipitacao_prevista}{data.table com a precipitacao prevista com as colunas
\itemize{
\item{data_rodada - data da rodada do modelo que gerou a previsao}
\item{data_previsao - data da previsao}
\item{cenario - codigo do cenario}
\item{nome - nome da sub bacia}
\item{valor - valor da previsao de precipitacao}
}}
}
\description{
Validador de arquivo de previsao de prec
}
