% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibracao.R
\name{calibracao}
\alias{calibracao}
\title{Funcao de calibracao do modelo SMAP/ONS}
\usage{
calibracao(
  vetor_modelo,
  kt_min,
  kt_max,
  area,
  EbInic,
  TuInic,
  Supin,
  precipitacao,
  evapotranspiracao,
  vazao,
  data_inicio_objetivo,
  data_fim_objetivo,
  limite_inferior,
  limite_superior,
  postos_plu,
  pesos = rep(1/length(vazao[, valor]), length(vazao[, valor]))
)
}
\arguments{
\item{vetor_modelo}{vetor resultante de unlist do objeto de classe smap_ons}

\item{kt_min}{valor do maximo lag maximo negativo}

\item{kt_max}{valor do maximo lag positivo}

\item{area}{area da sub-bacia}

\item{EbInic}{vazao de base inicial}

\item{TuInic}{umidade do solo inicial}

\item{Supin}{vazao superficial inicial}

\item{precipitacao}{data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{id - id do posto}
\item{valor - valor da variavel}
}}

\item{evapotranspiracao}{data table com a evapotranspiracao a ser ponderada com as colunas
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{id - id do posto}
\item{valor - valor da variavel}
}}

\item{vazao}{data table com a vazao a avaliada
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{id - id do posto}
\item{valor - valor da variavel}
}}

\item{data_inicio_objetivo}{data inicial da avaliacao da funcao objetivo}

\item{data_fim_objetivo}{data final da avaliacao da funcao objetivo}

\item{limite_inferior}{vetor com o limite inferior dos parametros}

\item{limite_superior}{vetor com o limite superior dos parametros}

\item{postos_plu}{data.table com as colunas
\itemize{
\item{nome - nome da sub_bacia}
\item{posto - nome do posto plu}
\item{valor - peso do posto plu}
}}

\item{pesos}{vetor de pesos a serem utilizados para cada data durante a calibracao}
}
\value{
ajuste valor da funcao objetivo
}
\description{
Raliza a calibracao de calibracao do SMAP/ONS dado um vetor inicial de pearametros e seus limites superiores e inferiores
}
