% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cria_inicializacao}
\alias{cria_inicializacao}
\title{Cria arquivo de inicializacao para uma unica sub-bacia}
\usage{
cria_inicializacao(
  parametros = NULL,
  nome = NULL,
  Ebin = NULL,
  Supin = NULL,
  Tuin = NULL,
  limite_inferior_ebin = 0.8,
  limite_superior_ebin = 1.2,
  limite_inferior_prec = 0.5,
  limite_superior_prec = 2,
  numero_dias_assimilacao = 32L,
  ajusta_precipitacao = 1L
)
}
\arguments{
\item{parametros}{data.table com as colunas
\itemize{
\item{nome - nome da sub-bacia (unico)}
\item{parametro - nome do parametro ("Ebin","Supin","Tuin")}
\item{valor - valor do parametro}
}}

\item{nome}{(opcional) nome da sub-bacia}

\item{Ebin}{valor para Ebin}

\item{Supin}{valor para Supin}

\item{Tuin}{valor para Tuin}

\item{limite_inferior_ebin}{vetor (length 1 ou 12) para limite inferior de Ebin}

\item{limite_superior_ebin}{vetor (length 1 ou 12) para limite superior de Ebin}

\item{limite_inferior_prec}{vetor (length 1 ou 12) para limite inferior de precipitacao}

\item{limite_superior_prec}{vetor (length 1 ou 12) para limite superior de precipitacao}

\item{numero_dias_assimilacao}{(integer) valor fixo (padrao = 32)}

\item{ajusta_precipitacao}{(integer) valor fixo (padrao = 1)}
}
\value{
data.table com colunas: nome, variavel, mes, valor
}
\description{
Cria arquivo de inicializacao para uma unica sub-bacia
}
