% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execucao.R
\name{executa_caso_oficial}
\alias{executa_caso_oficial}
\title{Executa caso oficial}
\usage{
executa_caso_oficial(pasta_entrada)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}
}
\value{
saida lista com o data.table previsao contendo as seguintes colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
otimizacao data table com as colunas:
\itemize{
\item{otimizacao}{valor das variaveis da otimizacao}
\item{nome}{nome da sub-bacia}
\item{data_caso}{data do caso}
}
assimilacao data table com as colunas:
\itemize{
\item{data_caso}{data da rodada}
\item{data_assimilacao}{data da assimilacao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
precipitacao data table contendo precipitacao observada e previsata com as colunas:
\itemize{
\item{data_previsao}{data da precipitacao}
\item{data_rodada}{data da rodada}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
}
\description{
executa rodada oficial do modleo SMAP/ONS
}
