% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilacao.R
\name{funcao_objetivo_assimilacao}
\alias{funcao_objetivo_assimilacao}
\title{funcao objetivo de assimilacao do SMAP/ONS}
\usage{
funcao_objetivo_assimilacao(
  vetorParametros,
  vetorModelo,
  TuInic,
  precipitacao_ponderada,
  evapotranspiracao,
  evapotranspiracao_planicie,
  vazao,
  area
)
}
\arguments{
\item{vetorParametros}{vetor com os parametros:
\itemize{
\item{pesos}{pesos da ponderacao da precipitacao}
\item{Ebin}{vazao de base inicial}
\item{Supin}{vazao superficial inicial}
}}

\item{vetorModelo}{vetor resultante de unlist do objeto de classe smap_ons}

\item{TuInic}{umidade do solo inicial}

\item{evapotranspiracao}{Vetor de ETo do final(ja corrigido e/ou ponderado)}

\item{vazao}{data table com a vazao a avaliada
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{area}{area da sub-bacia}

\item{pesos}{vetor de pesos a serem utilizados para a ponderacao da prec na assimilacao}

\item{EbInic}{vazao de base inicial}

\item{Supin}{vazao superficial inicial}

\item{precipitacao}{Vetor de precipitacao final (ja corrigido e/ou ponderado)}

\item{numero_dias}{numero de dias da assimilacao}
}
\value{
objetivo valor da funcao objetivo
}
\description{
funcao objetivo de assimilacao do SMAP/ONS
}
