% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibracao.R
\name{funcao_objetivo_calibracao}
\alias{funcao_objetivo_calibracao}
\title{funcao objetivo de calibracao do SMAP/ONS}
\usage{
funcao_objetivo_calibracao(
  vetor_modelo,
  kt_min,
  kt_max,
  area,
  EbInic,
  TuInic,
  Supin,
  precipitacao,
  evapotranspiracao,
  vazao,
  inicio_objetivo,
  fim_objetivo,
  postos_plu,
  pesos = rep(1/length(vazao[, valor]), length(vazao[, valor])),
  numero_dias,
  numero_postos_plu
)
}
\arguments{
\item{vetor_modelo}{vetor resultante de unlist do objeto de classe smap_ons}

\item{kt_min}{valor do maximo lag maximo negativo}

\item{kt_max}{valor do maximo lag positivo}

\item{area}{area da sub-bacia}

\item{EbInic}{vazao de base inicial}

\item{TuInic}{umidade do solo inicial}

\item{Supin}{vazao superficial inicial}

\item{precipitacao}{data table com a precipitacao a ser ponderada com as colunas
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{id - id do posto}
\item{valor - valor da variavel}
}}

\item{evapotranspiracao}{data table com a evapotranspiracao a ser ponderada com as colunas
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{id - id do posto}
\item{valor - valor da variavel}
}}

\item{vazao}{data table com a vazao a avaliada
\itemize{
\item{data - data da observacao}
\item{posto - nome do posto}
\item{id - id do posto}
\item{valor - valor da variavel}
}}

\item{inicio_objetivo}{inicio do calculo da funcao objetivo}

\item{fim_objetivo}{fim do calculo da funcao objetivo}

\item{postos_plu}{data.table com as colunas
\itemize{
\item{nome - nome da sub_bacia}
\item{posto - nome do posto plu}
\item{valor - peso do posto plu}
}}

\item{pesos}{vetor de pesos a serem utilizados para cada data durante a calibracao}

\item{numero_dias}{numero de dias da simulacao para a calibracao}

\item{numero_postos_plu}{numero de posto_plu considerados}
}
\value{
objetivo valor da funcao objetivo
}
\description{
Realiza o calculo da funcao objetivo para a calibracao do SMAP/ONS
}
\examples{
nome2 <- "baixoig"
 modelo <- new_modelo_smap_ons(parametros[nome == nome2], postos_plu[nome == nome2])
 kt_max <- sum(modelo$kt[1:2] > 0)
 kt_min <- sum(modelo$kt[4:63] > 0)

 EbInic <- 300
 TuInic <- 0.50
 Supin <- 700

 normal_climatologica <- historico_etp_NC[nome == nome2]
 precipitacao <- historico_precipitacao[posto \%in\% postos_plu[nome == nome2, posto]]
 precipitacao_ponderada <- ponderacao_espacial(precipitacao, postos_plu[nome == nome2])
 data_inicio_objetivo <- as.Date("2011-01-01")
 data_fim_objetivo <- as.Date("2021-12-31") - kt_max
 evapotranspiracao <- transforma_NC_serie(precipitacao_ponderada[data >= min(data) + kt_min & data <= data_fim_objetivo], normal_climatologica)
 vazao <- historico_vazao[data >= data_inicio_objetivo & data <= data_fim_objetivo & posto == nome2]

 area <- attributes(modelo)$area
 vetor_modelo <- unlist(modelo)
 vetor_modelo <- c(vetor_modelo[1:11], vetor_modelo[75:77], 5, 5)
 numero_postos_plu <- nrow(postos_plu[nome == nome2])

 limite_inferior <- vetor_modelo * 0.01
 limite_superior <- vetor_modelo * 10
 limite_inferior[8] <- 0.9999999
 limite_superior[8] <- 1.0000001
 limite_inferior[12] <- 0.8
 limite_superior[12] <- 1.2
 limite_inferior[13] <- 0.8
 limite_superior[13] <- 1.2
 limite_inferior[14] <- 0.8
 limite_superior[14] <- 1.2
 limite_inferior[15:16] <- 0.00001
 limite_superior[15:16] <- 50

 numero_dias <- nrow(evapotranspiracao)
 inicio_objetivo <- evapotranspiracao[data <= data_inicio_objetivo, .N]
 fim_objetivo <- evapotranspiracao[data <= data_fim_objetivo, .N]

\dontrun{
   fo <- funcao_objetivo_calibracao(vetor_modelo, kt_min, kt_max, area, EbInic, TuInic, Supin, precipitacao,
     evapotranspiracao, vazao, inicio_objetivo, fim_objetivo, postos_plu[nome == nome2],
     pesos = rep(1 / length(vazao[, valor]), length(vazao[, valor])), numero_dias, numero_postos_plu)
}
}
