% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_nova.R
\name{le_datas_rodada}
\alias{le_datas_rodada}
\title{Le csv com dados de datas de execucao do caso}
\usage{
le_datas_rodada(arq)
}
\arguments{
\item{arq}{nome do arquivo "inicializacao.csv"}
}
\value{
data table contendo as datas dos casos a serem executados e seus respectivos horizontes:
\itemize{
\item{data - data do caso}
\item{numero_dias_previsao - horizonte do caso}
}
}
\description{
Realiza a leitura do csv 'datasRodadas.csv' com as datas e o horizonte de previsao
}
