% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_inicializacao}
\alias{le_entrada_inicializacao}
\title{Leitor de arquivo de inicializacao do smap}
\usage{
le_entrada_inicializacao(pasta_entrada, nome_subbacia)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}
}
\value{
lista com o data.table inicializacao com as colunas
\itemize{
\item{nome - nome da sub-bacia}
\item{variavel - nome da variavel}
\item{valor - valor da variavel}
}
e data table datas_rodadas com as colunas
\itemize{
\item{data - data do caso}
\item{numero_dias_previsao - numero de dias de previsao de cada caso}
}
}
\description{
Le arquivo "sub-bacia_inicializacao.txt" utilizado no aplicativo SMAP/ONS
}
