% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_antiga.R
\name{le_entrada_parametros}
\alias{le_entrada_parametros}
\title{Leitor de arquivo de parametros}
\usage{
le_entrada_parametros(pasta_entrada, nome_subbacia)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{nome_subbacia}{nome da sub-bacia}
}
\value{
lista contendo os parametros da sub-bacia
}
\description{
Le arquivo de parametros "sub-bacia_PARAMETROS.txt".
}
