% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura_entrada.R
\name{le_entrada_previsao_precipitacao_0}
\alias{le_entrada_previsao_precipitacao_0}
\title{Leitor de arquivo de previsao de precipitacao do smap/ons}
\usage{
le_entrada_previsao_precipitacao_0(
  pasta_entrada,
  datas_rodadas,
  data_previsao,
  pontos_grade,
  nome_cenario
)
}
\arguments{
\item{pasta_entrada}{caminho da pasta  "arq_entrada"}

\item{datas_rodadas}{data table com as colunas
\itemize{
\item{data}{data da rodada}
\item{numero_dias_previsao}{numero de dias de previsao}
}}

\item{data_previsao}{data da previsao}

\item{pontos_grade}{data.table com as colunas
\itemize{
\item{nome}{nome da sub_bacia}
\item{nome_cenario_1}{primeira parte do nome dos cenarios de precipitacao considerados o caso}
\item{nome_cenario_2}{segunda parte do nome dos cenarios de precipitacao considerados o caso}
\item{latitude}{latitude do ponto de grade do cenario}
\item{longitude}{longitude do ponto de grade do cenario}
}}

\item{nome_cenario}{nome do cenario a ser simulado}
}
\value{
previsao_precipitacao data.table com as colunas
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{valor}{valor da previsao}
}
}
\description{
Le arquivo "'nome_cenario'\emph{p}'data_caso'a'data_previsao'.txt" utilizado no aplicativo SMAP/ONS
}
