% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_inicializacao}
\alias{le_inicializacao}
\title{Le csv com dados de inicializacao do caso}
\usage{
le_inicializacao(arq)
}
\arguments{
\item{arq}{nome do arquivo "inicializacao.csv"}
}
\value{
data.table com a inicializacao com as colunas
\itemize{
\item{nome - nome da sub_bacia}
\item{variavel - nome da variavel}
\item{valor - valor da variavel}
}
}
\description{
Realiza a leitura do csv 'inicializacao.csv' com os dados iniciais
}
