% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{le_parametros}
\alias{le_parametros}
\title{Leitor de arquivo de parametros}
\usage{
le_parametros(arq)
}
\arguments{
\item{arq}{o arquivo do tipo "subbacia_PARAMETROS.txt"}
}
\value{
data.table com as colunas
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da precipitacao observada}
}
}
\description{
Le arquivo de dados verificados, espera uma certa estrutura.
}
\details{
Esta funcao espera um arquivo da forma\tabular{lllll}{
   data \tab POSTO1 \tab POSTO2 \tab ... \tab POSTON \cr
   01/01/2020 \tab XXX \tab XXX \tab ... \tab XXX \cr
   02/01/2020 \tab XXX \tab XXX \tab ... \tab XXX \cr
   02/01/2020 \tab XXX \tab XXX \tab ... \tab XXX \cr
}


O nome da primeira coluna e irrelevante, pois sera trocado para "data" internamente.
Os demais serao utilizados como identificadores de cada registro
}
