% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_ons.R
\name{new_modelo_smap_ons}
\alias{new_modelo_smap_ons}
\title{Construtor do modelo classe \code{smap_ons}}
\usage{
new_modelo_smap_ons(parametros)
}
\arguments{
\item{parametros}{data table com 8910 linhas and 3 colunas:
\describe{
\itemize{
\item{nome}{nome da sub-bacia}
\item{parametros}{nome do parametros}
\item{valor}{valor do parametro}
}
}}
}
\description{
Construtor do modelo classe \code{smap_ons}
}
