% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parametros}
\alias{parametros}
\title{Parametros}
\format{
\subsection{\code{parametros}}{

A data table com 8910 linhas and 3 colunas:
\describe{
\itemize{
\item{nome}{nome da sub-bacia}
\item{parametros}{nome do parametros}
\item{valor}{valor do parametro}
}
}
}
}
\source{
ONS
}
\usage{
parametros
}
\description{
Parametros das sub-bacias modeladas pelo SMAP/ONS
}
\keyword{datasets}
