% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{ponderacao_temporal}
\alias{ponderacao_temporal}
\title{Funcao de ponderacao temporal}
\usage{
ponderacao_temporal(serie_temporal, kt, kt_max, kt_min)
}
\arguments{
\item{serie_temporal}{vetor com a variavel a ser ponderada}

\item{kt}{vetor de kts}

\item{kt_max}{valor do maximo lag positivo}

\item{kt_min}{valor do maximo lag maximo negativo}
}
\value{
serie_temporal_ponderada: vetor com a serie temporal ponderada pelos kts
}
\description{
Realiza a ponderacao temporal de variaveis atraves dos pesos kt
}
