% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderacao.R
\name{ponderacao_temporal}
\alias{ponderacao_temporal}
\title{Funcao de ponderacao temporal}
\usage{
ponderacao_temporal(serie_temporal, kt, kt_max, kt_min)
}
\arguments{
\item{serie_temporal}{vetor com a variavel a ser ponderada}

\item{kt}{vetor de kts}

\item{kt_max}{valor do maximo lag positivo}

\item{kt_min}{valor do maximo lag maximo negativo}
}
\value{
serie_temporal_ponderada: vetor com a serie temporal ponderada pelos kts
}
\description{
Realiza a ponderacao temporal de variaveis atraves dos pesos kt
}
\examples{
sub_bacia <- "avermelha"
modelo <- new_modelo_smap_ons(parametros[nome == sub_bacia], postos_plu[nome == sub_bacia])
kt <- unlist(modelo)[12:74]
kt_min <- sum(kt[4:63] > 0)
kt_max <- sum(kt[1:2] > 0)
data_inicio <- as.Date("2021-12-01")
data_fim <- as.Date("2021-12-29")

serie_temporal <- historico_precipitacao[posto == "psatagv" &  data >= data_inicio - kt_min & data <= data_fim + kt_max, valor]

saida <- ponderacao_temporal(serie_temporal, kt, kt_max, kt_min)
}
