% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodadas_encadeadas.R
\name{rodada_encadeada_oficial}
\alias{rodada_encadeada_oficial}
\title{Funcao de calibracao do modelo SMAP/ONS}
\usage{
rodada_encadeada_oficial(
  parametros,
  inicializacao,
  historico_precipitacao,
  previsao_precipitacao,
  historico_etp_NC,
  historico_vazao,
  postos_plu,
  datas_rodadas,
  numero_dias_assimilacao,
  numero_cenarios,
  sub_bacias
)
}
\arguments{
\item{parametros}{data table com os parametros dos modelos
\itemize{
\item{Nome}{Nome da sub-bacia}
\item{parametro}{nome do parametro}
\item{valor}{valor do parametro}
}}

\item{inicializacao}{data.table com a inicializacao com as colunas:
\itemize{
\item{nome}{nome da sub_bacia}
\item{Ebin}{vazao de base inicial}
\item{Supin}{vazao superficial inicial}
\item{Yuin}{umidade do solo inicial}
}}

\item{historico_precipitacao}{data table com o historico de precipitacao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{previsao_precipitacao}{data.table com a previsao de precipitacao com as colunas:
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da previsao}
}}

\item{historico_etp_NC}{data.table com o historico de NC deevapotranspiracao com as colunas:
\itemize{
\item{mes}{mes da NC}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da NC de evapotranspiracao observada}
}}

\item{historico_vazao}{data table com o historico de vazao com as colunas:
\itemize{
\item{data}{data da observacao}
\item{posto}{nome do posto}
\item{id}{id do posto}
\item{valor}{valor da variavel}
}}

\item{postos_plu}{data table contendo a relacao sub-bacia x postos_plu com as colunas:
\itemize{
\item{nome}{nome da sub_bacia}
\item{posto}{nome do posto plu}
\item{valor}{peso do posto plu}
}}

\item{datas_rodadas}{data table contendo as datas dos casos a serem executados e seus respectivos horizontes:
\itemize{
\item{data}{data do caso}
\item{numero_dias_previsao}{horizonte do caso}
}}

\item{numero_dias_assimilacao}{numero de dias para o processo de assimilação de dados}

\item{numero_cenarios}{numero de cenarios a serem gerados}

\item{sub_bacias}{vetor com o nome das sub-bacias a serem consideradas}
}
\value{
saida data table contendo as seguintes colunas:
\itemize{
\item{data_rodada}{data da rodada}
\item{data_previsao}{data da previsao}
\item{cenario}{nome do cenario}
\item{nome}{nome da sub-bacia}
\item{variavel}{nome da variavel}
\item{valor}{valor da variavel}
}
}
\description{
Raliza a calibracao de calibracao do SMAP/ONS dado um vetor inicial de pearametros e seus limites superiores e inferiores
}
