% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_ons.R
\name{smap_ons.previsao}
\alias{smap_ons.previsao}
\title{Rodada de um dia do SMAP}
\usage{
smap_ons.previsao(
  parametros,
  inicializacao,
  precipitacao,
  evapotranspiracao,
  Emarg,
  saidaAnterior = matrix(nrow = 0, ncol = length(matrizSaida_nom)),
  dataD = NULL
)
}
\arguments{
\item{parametros}{data table com 8910 linhas and 3 colunas:
\describe{
\itemize{
\item{nome}{nome da sub-bacia}
\item{parametros}{nome do parametros}
\item{valor}{valor do parametro}
}
}}

\item{inicializacao}{Objeto resultante da funcao smap.inicializacao}

\item{precipitacao}{Valor de precipitacao final do dia (ja corrigido e/ou ponderado)}

\item{evapotranspiracao}{Valor de ETo do final dia (ja corrigido e/ou ponderado)}

\item{Emarg}{Valor da evaporacao final do dia do reservatorio de planicie (ja corrigido e/ou ponderado)}

\item{saidaAnterior}{matriz de saida da iteracao passada (primeira iteracao nao atribuir nenhum valor)}

\item{dataD}{Nomea a linha da matriz de saida com a data fornecida}
}
\value{
Matriz com a vazao calculada e os valores de estado e funcoes de transferencia
}
\description{
Rodada de um dia do SMAP
}
